+++ .enzymeSetup.ts
@@ -1,4 +1,5 @@
 import * as Enzyme from "enzyme"
 import Adapter from "enzyme-adapter-react-16"
 
+// Todo: what is this doing? Do we need it?
 Enzyme.configure({ adapter: new Adapter() })
