+++ charts/core/ChartUrl.test.ts
@@ -2,7 +2,7 @@
 
 import { ChartConfigProps } from "charts/core/ChartConfig"
 import { TimeBoundValue, TimeBound, TimeBounds } from "charts/utils/TimeBounds"
-import { createConfig, setupChart } from "test/utils"
+import { createConfig, setupChart } from "charts/test/utils"
 
 import { ChartUrl, ChartQueryParams, EntityUrlBuilder } from "./ChartUrl"
 import { MapConfigProps } from "charts/mapCharts/MapConfig"
