+++ charts/readme.md
@@ -1,3 +1,3 @@
 # Charts
 
-This folder contains the code for our client side chart library. 
+This folder contains the code for our client side chart library.
