+++ owidTable/OwidTable.test.ts
@@ -1,7 +1,7 @@
 #! /usr/bin/env yarn jest
 
 import { OwidTable, BasicTable } from "owidTable/OwidTable"
-import { readVariable } from "test/fixtures"
+import { readVariable } from "charts/test/fixtures"
 
 describe(OwidTable, () => {
     // Scenarios
