+++ utils/styleMock.ts
@@ -1,0 +1,3 @@
+// Jest cannot handle importing CSS so we have CSS imports mapped to this empty object
+// https://stackoverflow.com/questions/39418555/syntaxerror-with-jest-and-react-and-importing-css-files
+export default {}
